pyproject.toml
[project]
name = "turbo-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.14.*"
dependencies = [
    "bytecode>=0.17.0",
    "pyturso>=0.5.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
